requirements.txt
Django==1.6
argparse==1.2.1
dj-database-url==0.2.2
dj-static==0.0.5
django-toolbelt==0.0.1
gunicorn==18.0
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
django-filter==0.7
django-haystack==2.1.0
djangorestframework==2.3.9
